restrict access to firestore based on user id

+++ functions/src/domains/v1/contacts/firestore-repository.ts
@@ -8,35 +8,57 @@ import {
 const contactsCollection = db.collection("contacts");
 
 const FirestoreRepository: IContactRepository = {
-  async getAllContacts(): Promise<IContact[]> {
-    const snapshot = await contactsCollection.get();
+  async getAllContacts(userId: string): Promise<IContact[]> {
+    const snapshot = await contactsCollection
+      .where("ownerId", "==", userId)
+      .get();
     return snapshot.docs.map(
       (doc) => ({ id: doc.id, ...doc.data() }) as IContact,
     );
   },
 
-  async getContactById(id: string): Promise<IContact | null> {
+  async getContactById(userId: string, id: string): Promise<IContact | null> {
     const doc = await contactsCollection.doc(id).get();
-    return doc.exists ? ({ id: doc.id, ...doc.data() } as IContact) : null;
+    return doc.exists && doc.data()?.ownerId === userId
+      ? ({ id: doc.id, ...doc.data() } as IContact)
+      : null;
   },
 
-  async addContact(contact: ContactWithoutId): Promise<IContact> {
-    const docRef = await contactsCollection.add(contact);
-    return { id: docRef.id, ...contact };
+  async addContact(
+    userId: string,
+    contact: ContactWithoutId,
+  ): Promise<IContact> {
+    const docRef = await contactsCollection.add({
+      ...contact,
+      ownerId: userId,
+    });
+    const doc = await docRef.get();
+    return { id: docRef.id, ...doc.data() } as IContact;
   },
 
   async updateContact(
+    userId: string,
     id: string,
     contact: Partial<IContact>,
   ): Promise<IContact | null> {
     const docRef = contactsCollection.doc(id);
-    await docRef.update(contact);
-    return this.getContactById(id);
+    const doc = await docRef.get();
+    if (doc.exists && doc.data()?.ownerId === userId) {
+      await docRef.update(contact);
+      return this.getContactById(userId, id);
+    }
+    return null;
   },
 
-  async deleteContact(id: string): Promise<void> {
+  async deleteContact(userId: string, id: string): Promise<IContact | null> {
     const docRef = contactsCollection.doc(id);
-    await docRef.delete();
+    const doc = await docRef.get();
+    if (doc.exists && doc.data()?.ownerId === userId) {
+      const contact = { id: doc.id, ...doc.data() } as IContact;
+      await docRef.delete();
+      return contact; // Return the contact data being removed
+    }
+    return null; // Return null if the contact is not found
   },
 };
